requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
distro==1.9.0
google-api-core==2.25.1
google-api-python-client==2.108.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
nexus-rpc==1.1.0
oauthlib==3.3.1
openai==1.30.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
slack_sdk==3.27.0
sniffio==1.3.1
temporalio==1.14.1
tqdm==4.67.1
types-protobuf==6.30.2.20250703
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
